+++ src/frontend/apps/lti_site/components/App/AppContentLoader/index.spec.tsx
@@ -4,7 +4,6 @@ import {
   AppConfigProvider,
   appNames,
   appState,
-  decodeJwt,
   modelName,
   useCurrentSession,
   useCurrentUser,
@@ -15,12 +14,6 @@ import React from 'react';
 import { IntlShape, useIntl } from 'react-intl';
 
 import AppContentLoader from '.';
-
-jest.mock('lib-components', () => ({
-  ...jest.requireActual('lib-components'),
-  decodeJwt: jest.fn(),
-}));
-const mockedDecodeJwt = decodeJwt as jest.MockedFunction<typeof decodeJwt>;
 
 let intl: Maybe<IntlShape>;
 const MockContent = () => {
@@ -34,29 +27,14 @@ describe('<AppContentLoader />', () => {
   beforeEach(() => {
     useCurrentUser.setState({ currentUser: undefined });
     useCurrentSession.setState({ sessionId: undefined });
-
     intl = undefined;
+
+    jest.resetAllMocks();
   });
 
   it('initialize current user before loading content', async () => {
-    useJwt.setState({ jwt: 'some jwt' });
-    mockedDecodeJwt.mockReturnValue({
-      session_id: 'some session id',
-      user: {
-        anonymous_id: 'anonymous id',
-        email: 'some email',
-        id: 'id',
-        username: 'user name',
-        user_fullname: 'user full name',
-      },
-      locale: 'pl',
-      maintenance: false,
-      permissions: {
-        can_access_dashboard: false,
-        can_update: false,
-      },
-      resource_id: 'resource id',
-      roles: [],
+    useJwt.setState({
+      jwt: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzZXNzaW9uX2lkIjoic29tZV9zZXNzaW9uX2lkIiwidXNlciI6eyJhbm9ueW1vdXNfaWQiOiJhbm9ueW1vdXMgaWQiLCJlbWFpbCI6InNvbWUgZW1haWwiLCJpZCI6ImlkIiwidXNlcm5hbWUiOiJ1c2VyIG5hbWUiLCJ1c2VyX2Z1bGxuYW1lIjoidXNlciBmdWxsIG5hbWUifSwibG9jYWxlIjoicGwiLCJtYWludGVuYW5jZSI6ZmFsc2UsInBlcm1pc3Npb25zIjp7ImNhbl9hY2Nlc3NfZGFzaGJvYXJkIjpmYWxzZSwiY2FuX3VwZGF0ZSI6ZmFsc2V9LCJyZXNvdXJjZV9pZCI6InJlc291cmNlIGlkIiwicm9sZXMiOltdfQ.gv0kmitQfOv93TQuFTHsiqQJFWeTkbmb1h8J8uMVX70',
     });
 
     render(
@@ -91,9 +69,8 @@ describe('<AppContentLoader />', () => {
 
     await screen.findByText('content');
 
-    expect(decodeJwt).toHaveBeenCalledTimes(1);
     expect(useCurrentSession.getState()).toEqual({
-      sessionId: 'some session id',
+      sessionId: 'some_session_id',
     });
     expect(useCurrentUser.getState().currentUser).toEqual({
       anonymous_id: 'anonymous id',
@@ -107,5 +84,41 @@ describe('<AppContentLoader />', () => {
     });
 
     expect(intl?.locale).toEqual('pl');
+  });
+
+  it('renders an error page if no jwt provided', async () => {
+    render(
+      <AppConfigProvider
+        value={{
+          appName: appNames.CLASSROOM,
+          attendanceDelay: 10,
+          state: appState.SUCCESS,
+          modelName: modelName.VIDEOS,
+          sentry_dsn: 'test.dns.com',
+          environment: 'tests',
+          frontend: 'test-frontend',
+          release: 'debug',
+          static: {
+            svg: {
+              icons: '',
+            },
+            img: {
+              liveBackground: '',
+              liveErrorBackground: '',
+              marshaWhiteLogo: '',
+              videoWizardBackground: '',
+              errorMain: '',
+            },
+          },
+          uploadPollInterval: 10,
+        }}
+      >
+        <AppContentLoader />
+      </AppConfigProvider>,
+    );
+
+    screen.queryByText(
+      'Unable to find a jwt Token. The ressource might not exist.',
+    );
   });
 });