+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Bug on chat messages synchronisation
 - Static files exclusion
+- Impossible to decode JWT token, there is no jwt to decode
 
 ### Removed